requirements.txt
gradio>=4.0.0
torch>=2.0.0
transformers>=4.30.0
pillow>=9.0.0
numpy>=1.24.0 